Extend post-firmware DOM verification to threshold validation
